subscriber/requirements.txt
azure-common==1.1.27
azure-core==1.15.0
azure-servicebus==7.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dataclasses==0.8
Flask==2.0.1
idna==2.10
importlib-metadata==4.5.0
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
msrest==0.6.21
oauthlib==3.1.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
typing-extensions==3.10.0.0
uamqp==1.4.0
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.4.1
